# Application/DTOs/Countries/GetCountriesNamesRequestDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Application.DTOs.Countries
{
    public class GetCountriesNamesRequestDto
    {
        public string Name { get; set; }
    }
}