# Domain/Entities/Alerts.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Domain.Entities
{
    public class Alerts
    {
        [Key]
        public int AlertId { get; set; }
        public string Content { get; set; }
        public DateTime CreatedDate { get; set; }
        public DateTime? ValidDate { get; set; }
        public int CountryId { get; set; }
    }
}